the fix

+++ 0x1E-search_algorithms/1-binary.c
@@ -28,7 +28,7 @@ int binary_search(int *array, size_t size, int value)
 	else if (array[mid] > value)
 	right = mid - 1;
 	else
-	return ((int)mid);
+	return (mid);
 	}
 	return (-1);
 }
